requirements.txt
# JSON schema validation for game data
jsonschema>=4.17.0

# Terminal output coloring
termcolor>=2.3.0

# Progress bars for long operations
tqdm>=4.65.0

# HTTP requests for web interactions
requests>=2.31.0

# Token counting for OpenAI API usage tracking
tiktoken>=0.5.0

# Image processing for JPEG compression and thumbnail generation
Pillow>=10.0.0

# Multi-LLM provider support
google-generativeai>=0.3.0  # For Gemini API support
ollama>=0.1.0                # For Ollama local model support

# Proxy server for LM Studio mode and API call capture
mitmproxy>=10.0.0

# Web interface
flask>=2.3.0
flask-socketio>=5.3.0
python-socketio>=5.9.0
python-engineio>=4.7.0

# Python standard library extensions (usually included)
# pathlib - included in Python 3.4+
# datetime - standard library
# json - standard library
# os - standard library
# sys - standard library
# re - standard library
# subprocess - standard library
# threading - standard library
# queue - standard library
# time - standard library
# io - standard library
# contextlib - standard library
# shutil - standard library
# codecs - standard library

# Optional: For enhanced development experience
# black>=23.0.0          # Code formatting
# flake8>=6.0.0          # Code linting
# pytest>=7.0.0          # Testing framework
# mypy>=1.0.0            # Type checking

# Note: Make sure you have Python 3.9+ installed
# Install with: pip install -r requirements.txt